Fixes rendering animation when only armature (not mesh) is having animation assigned
For common case when mesh is parented to armature and armature is animated. Note that in example Blue.blend both
armature and mesh are animated so animation is displayed correctly. However, in common case when
only armature is added to NLA editor or dopesheet this was not OK. Following code checks if object
has animation, if not checks parent and if armature and if animated, it uses that one. So, this way it works like in Blender.